refactor:  move should_enable_page_index from mod.rs to opener.rs
